# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deni",
        "description": "Descripción de la API de Ejemplo",
        "termsOfService": "https://www.ejemplo.com/terms",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "license": {
            "name": "Licencia de Ejemplo",
            "url": "https://www.ejemplo.com/licencia"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/estados-evento": {
            "get": {
                "tags": [
                    "Estados de Evento"
                ],
                "summary": "Obtener todos los estados de eventos",
                "description": "Recupera todos los estados de eventos de la base de datos.",
                "operationId": "indexEstEv",
                "responses": {
                    "200": {
                        "description": "Lista de estados de eventos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/estado_evento"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Estados de Evento"
                ],
                "summary": "Registrar un nuevo estado de evento",
                "description": "Crear un nuevo estado de evento con la información proporcionada.",
                "operationId": "storeEstEv",
                "requestBody": {
                    "description": "Datos del estado de evento",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nombre del Estado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Descripción del Estado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estado de evento registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/estado_evento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/estados-evento/{estado}": {
            "get": {
                "tags": [
                    "Estados de Evento"
                ],
                "summary": "Obtener información de un estado de evento",
                "description": "Obtiene información de un estado de evento por su ID.",
                "operationId": "showEstEv",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "path",
                        "description": "ID del estado de evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del estado de evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/estado_evento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estado de evento no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Estado de evento no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Estados de Evento"
                ],
                "summary": "Actualizar un estado de evento",
                "description": "Actualiza un estado de evento existente con la información proporcionada.",
                "operationId": "updateEstEv",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "path",
                        "description": "ID del estado de evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos actualizados del estado de evento",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nuevo Nombre del Estado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nueva Descripción del Estado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado de evento actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/estado_evento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estado de evento no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Estado de evento no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estados de Evento"
                ],
                "summary": "Eliminar un estado de evento",
                "description": "Elimina un estado de evento existente por su ID.",
                "operationId": "destroyEstEv",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "path",
                        "description": "ID del estado de evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Estado de evento eliminado exitosamente"
                    },
                    "404": {
                        "description": "Estado de evento no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Estado de evento no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "estado_evento": {
                "title": "estado_evento",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}